refactor : 구분자 관련 함수 클래스로 분리
