+++ src/main/kotlin/calculator/BasicCalculator.kt
@@ -1,0 +1,7 @@
+package calculator
+
+class BasicCalculator {
+    fun add(input: List<String>): Int {
+        return input.sumOf { it.toInt() }
+    }
+}
